Update HTML links for relative paths and refactor flow.js functions
