+++ index.html
@@ -21,7 +21,7 @@
   <body>
     <div class="sidebar">
       <ul>
-        <li><a href="/timeline.html">Timeline</a></li>
+        <li><a href="./timeline.html">Timeline</a></li>
         <li onclick="clearData()" id="cleardata">Clear Data</li>
       </ul>
     </div>
@@ -64,7 +64,7 @@
     </div>
 
     <ul class="timeline" id="listTimeline">
-      <a href="timeline.html" class="view-timeline"></a>
+      <a href="./timeline.html" class="view-timeline"></a>
     </ul>
 
     <div class="charts">
@@ -78,6 +78,6 @@
         <canvas id="chartSavings"></canvas>
       </div>
     </div>
-    <script type="module" src="/src/main.js"></script>
+    <script type="module" src="./src/main.js"></script>
   </body>
 </html>
