+++ timeline.html
@@ -22,7 +22,7 @@
   <body>
     <div class="sidebar">
       <ul>
-        <li><a href="/">Home</a></li>
+        <li><a href="./index.html">Home</a></li>
         <li id="cleardata">Clear Data</li>
       </ul>
     </div>
